fixed percentage hit only being 0 or 100

+++ SpaceTurtleSpeedTyping/GameLib/stats.cpp
@@ -47,7 +47,7 @@ double Stats::getCorrectRate()
     {
         return 0.0;
     } else {
-        return (correctTypeCount / totalTypeCount) * 100;
+        return ((double)correctTypeCount / (double)totalTypeCount) * 100;
     }
 }
